% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonClassify}
\alias{GNNMonClassify}
\title{Classify for "membership".}
\usage{
GNNMonClassify(gnnObj, points = NULL)
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{points}{A numeric vector (a point) or a matrix or a data frame
of points. If NULL monad's data is used.}

\item{method}{A string for the nearest neighbors finding method.
One of "scan", or  "kdtree".}
}
\value{
A GNNMon object
}
\description{
Classifies a given point as close enough to
the monad points (TRUE) or too distant (FALSE).
}
\details{
If points is a matrix or a data frame then its number of columns
should equal the dimension of monad's points.
The result is assigned to \code{gnnObj$Value}.
}
