% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonComputeNearestNeighborDistances}
\alias{GNNMonComputeNearestNeighborDistances}
\title{Compute nearest neighbors distances.}
\usage{
GNNMonComputeNearestNeighborDistances(
  gnnObj,
  nTopNNs = 6,
  radius = Inf,
  distanceMethod = "euclidean"
)
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{nTopNNs}{Number of top nearest neighbors.}

\item{radius}{Radius within which the nearest neighbors are taken.}

\item{distanceMethod}{A string for the distance method.
One of "euclidean", "cosine".}
}
\value{
A GNNMon object
}
\description{
Computes the nearest neighbors and corresponding distances
using a specified distance function for each row of monad's data matrix.
}
