% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonComputeProximityMatrix}
\alias{GNNMonComputeProximityMatrix}
\title{Compute proximity matrix.}
\usage{
GNNMonComputeProximityMatrix(gnnObj, n = NULL)
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{n}{Number of nearest neighbors.}
}
\value{
A GNNMon object
}
\description{
Computes the proximity (similarity) matrix using the matrix
\code{gnnObj$NearestNeighborDistances}.
}
\details{
The result is assigned to \code{gnnObj$Value}.
}
