% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonCreateKDTreeObject}
\alias{GNNMonCreateKDTreeObject}
\title{Create KDTreeObject.}
\usage{
GNNMonCreateKDTreeObject(
  gnnObj,
  distanceFunction = "euclidean",
  functionName = "GNNMonCreateKDTreeObject"
)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{distanceFunction}{Distance function spec.
One of NULL, "euclidean", "cosine".}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{GNNMonFailureSymbol}.

A GNNMon object
}
\description{
Creates KDTreeObject from the monad object's data and distance function.
}
