% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonEchoFunctionValue}
\alias{GNNMonEchoFunctionValue}
\title{Echo function application to monad's value.}
\usage{
GNNMonEchoFunctionValue(gnnObj, f)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{f}{A function to be applied to \code{gnnObj$Value}.}
}
\value{
A GNNMon object.
}
\description{
Applies a function to the "Value" element/member of the monad object
and prints the result.
}
\details{
Prints \code{f(gnnObj$Value)}.
}
