% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonEchoValue}
\alias{GNNMonEchoValue}
\title{Echo monad's value.}
\usage{
GNNMonEchoValue(gnnObj)
}
\arguments{
\item{gnnObj}{An GNNMon object.}
}
\value{
A GNNMon object.
}
\description{
Prints the "Value" element/member of the monad object.
}
\details{
Prints \code{f(gnnObj$Value)}.
}
