% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonCore.R
\docType{data}
\name{GNNMonFailureSymbol}
\alias{GNNMonFailureSymbol}
\title{Failure symbol for GNNMon.}
\format{An object of class \code{logical} of length 1.}
\usage{
GNNMonFailureSymbol
}
\description{
Failure symbol for the monad GNNMon.
}
\keyword{datasets}
