% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonFindNearest}
\alias{GNNMonFindNearest}
\title{Find nearest points.}
\usage{
GNNMonFindNearest(gnnObj, point, n = 12, method = "euclidean")
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{point}{A numeric vector.}

\item{method}{A string for the distance method.
One of "euclidean", "cosine".}
}
\value{
A GNNMon object
}
\description{
Finds the nearest points from the monad to a given point.
}
\details{
The result is assigned to \code{gnnObj$Value}.
This is a fairly brute force algorithm because in order to find
the top nearest neighbors the distances to all monad points are computed.
}
