% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonSetDistanceMethod}
\alias{GNNMonSetDistanceMethod}
\title{Set DistanceMethod.}
\usage{
GNNMonSetDistanceMethod(gnnObj, DistanceMethod)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{DistanceMethod}{An object member to be set.}
}
\value{
An GNNMon object.
}
\description{
Sets DistanceMethod into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{GNNMonSetData}()},
\code{\link{GNNMonSetLowerThreshold}()},
\code{\link{GNNMonSetNearestNeighborDistances}()},
\code{\link{GNNMonSetNumberOfNNs}()},
\code{\link{GNNMonSetRadiusFunction}()},
\code{\link{GNNMonSetRadius}()},
\code{\link{GNNMonSetUpperThreshold}()},
\code{\link{GNNMonSetValue}()},
\code{\link{GNNMonTakeData}()},
\code{\link{GNNMonTakeDistanceMethod}()},
\code{\link{GNNMonTakeLowerThreshold}()},
\code{\link{GNNMonTakeNearestNeighborDistances}()},
\code{\link{GNNMonTakeNumberOfNNs}()},
\code{\link{GNNMonTakeRadiusFunction}()},
\code{\link{GNNMonTakeRadius}()},
\code{\link{GNNMonTakeUpperThreshold}()},
\code{\link{GNNMonTakeValue}()}
}
\concept{Set/Take functions}
