% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonSetKDTreeObject}
\alias{GNNMonSetKDTreeObject}
\title{Set KDTreeObject.}
\usage{
GNNMonSetKDTreeObject(gnnObj, KDTreeObject)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{KDTreeObject}{An object member to be set.}
}
\value{
An GNNMon object.
}
\description{
Sets KDTreeObject into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{GNNMonSetData}()},
\code{\link{GNNMonSetDistanceFunction}()},
\code{\link{GNNMonSetLowerThreshold}()},
\code{\link{GNNMonSetNearestNeighborDistances}()},
\code{\link{GNNMonSetNumberOfNNs}()},
\code{\link{GNNMonSetRadius}()},
\code{\link{GNNMonSetRadiusFunction}()},
\code{\link{GNNMonSetUpperThreshold}()},
\code{\link{GNNMonSetValue}()},
\code{\link{GNNMonTakeData}()},
\code{\link{GNNMonTakeDistanceFunction}()},
\code{\link{GNNMonTakeKDTreeObject}()},
\code{\link{GNNMonTakeLowerThreshold}()},
\code{\link{GNNMonTakeNearestNeighborDistances}()},
\code{\link{GNNMonTakeNumberOfNNs}()},
\code{\link{GNNMonTakeRadius}()},
\code{\link{GNNMonTakeRadiusFunction}()},
\code{\link{GNNMonTakeUpperThreshold}()},
\code{\link{GNNMonTakeValue}()}
}
\concept{Set/Take functions}
