% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonCore.R
\name{GNNMonSetLowerThreshold}
\alias{GNNMonSetLowerThreshold}
\title{Set LowerThreshold.}
\usage{
GNNMonSetLowerThreshold(gnnObj, LowerThreshold)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{LowerThreshold}{An object member to be set.}
}
\value{
An GNNMon object.
}
\description{
Sets LowerThreshold into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{GNNMonSetData}},
  \code{\link{GNNMonSetDistanceMethod}},
  \code{\link{GNNMonSetNumberOfNNs}},
  \code{\link{GNNMonSetRadiusFunction}},
  \code{\link{GNNMonSetRadius}},
  \code{\link{GNNMonSetUpperThreshold}},
  \code{\link{GNNMonSetValue}},
  \code{\link{GNNMonTakeData}},
  \code{\link{GNNMonTakeDistanceMethod}},
  \code{\link{GNNMonTakeLowerThreshold}},
  \code{\link{GNNMonTakeNumberOfNNs}},
  \code{\link{GNNMonTakeRadiusFunction}},
  \code{\link{GNNMonTakeRadius}},
  \code{\link{GNNMonTakeUpperThreshold}},
  \code{\link{GNNMonTakeValue}}
}
\concept{Set/Take functions}
