% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonSetRadius}
\alias{GNNMonSetRadius}
\title{Set Radius.}
\usage{
GNNMonSetRadius(gnnObj, Radius)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{Radius}{An object member to be set.}
}
\value{
An GNNMon object.
}
\description{
Sets Radius into the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{GNNMonSetData}()},
\code{\link{GNNMonSetDistanceMethod}()},
\code{\link{GNNMonSetKDTreeObject}()},
\code{\link{GNNMonSetLowerThreshold}()},
\code{\link{GNNMonSetNearestNeighborDistances}()},
\code{\link{GNNMonSetNumberOfNNs}()},
\code{\link{GNNMonSetRadiusFunction}()},
\code{\link{GNNMonSetUpperThreshold}()},
\code{\link{GNNMonSetValue}()},
\code{\link{GNNMonTakeData}()},
\code{\link{GNNMonTakeDistanceMethod}()},
\code{\link{GNNMonTakeKDTreeObject}()},
\code{\link{GNNMonTakeLowerThreshold}()},
\code{\link{GNNMonTakeNearestNeighborDistances}()},
\code{\link{GNNMonTakeNumberOfNNs}()},
\code{\link{GNNMonTakeRadius}()},
\code{\link{GNNMonTakeRadiusFunction}()},
\code{\link{GNNMonTakeUpperThreshold}()},
\code{\link{GNNMonTakeValue}()}
}
\concept{Set/Take functions}
