% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonSetUpperThreshold}
\alias{GNNMonSetUpperThreshold}
\title{Set UpperThreshold.}
\usage{
GNNMonSetUpperThreshold(gnnObj, UpperThreshold)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{UpperThreshold}{An object member to be set.}
}
\value{
An GNNMon object.
}
\description{
Sets UpperThreshold into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{GNNMonSetData}},
  \code{\link{GNNMonSetDistanceMethod}},
  \code{\link{GNNMonSetLowerThreshold}},
  \code{\link{GNNMonSetNearestNeighborDistances}},
  \code{\link{GNNMonSetNumberOfNNs}},
  \code{\link{GNNMonSetRadiusFunction}},
  \code{\link{GNNMonSetRadius}},
  \code{\link{GNNMonSetValue}},
  \code{\link{GNNMonTakeData}},
  \code{\link{GNNMonTakeDistanceMethod}},
  \code{\link{GNNMonTakeLowerThreshold}},
  \code{\link{GNNMonTakeNearestNeighborDistances}},
  \code{\link{GNNMonTakeNumberOfNNs}},
  \code{\link{GNNMonTakeRadiusFunction}},
  \code{\link{GNNMonTakeRadius}},
  \code{\link{GNNMonTakeUpperThreshold}},
  \code{\link{GNNMonTakeValue}}
}
\concept{Set/Take functions}
