% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonTakeData}
\alias{GNNMonTakeData}
\title{Take Data.}
\usage{
GNNMonTakeData(gnnObj, functionName = "GNNMonTakeData")
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{GNNMonFailureSymbol}.
}
\description{
Takes Data from the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{GNNMonSetData}},
  \code{\link{GNNMonSetDistanceMethod}},
  \code{\link{GNNMonSetLowerThreshold}},
  \code{\link{GNNMonSetNumberOfNNs}},
  \code{\link{GNNMonSetRadiusFunction}},
  \code{\link{GNNMonSetRadius}},
  \code{\link{GNNMonSetUpperThreshold}},
  \code{\link{GNNMonSetValue}},
  \code{\link{GNNMonTakeDistanceMethod}},
  \code{\link{GNNMonTakeLowerThreshold}},
  \code{\link{GNNMonTakeNumberOfNNs}},
  \code{\link{GNNMonTakeRadiusFunction}},
  \code{\link{GNNMonTakeRadius}},
  \code{\link{GNNMonTakeUpperThreshold}},
  \code{\link{GNNMonTakeValue}}
}
\concept{Set/Take functions}
