% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonTakeLowerThreshold}
\alias{GNNMonTakeLowerThreshold}
\title{Take LowerThreshold.}
\usage{
GNNMonTakeLowerThreshold(gnnObj, functionName = "GNNMonTakeLowerThreshold")
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{functionName}{A string that is a name of this function or a delegating function.}
}
\value{
A list of functions or \code{GNNMonFailureSymbol}.
}
\description{
Takes LowerThreshold from the monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{GNNMonSetData}()},
\code{\link{GNNMonSetDistanceFunction}()},
\code{\link{GNNMonSetKDTreeObject}()},
\code{\link{GNNMonSetLowerThreshold}()},
\code{\link{GNNMonSetNearestNeighborDistances}()},
\code{\link{GNNMonSetNumberOfNNs}()},
\code{\link{GNNMonSetRadius}()},
\code{\link{GNNMonSetRadiusFunction}()},
\code{\link{GNNMonSetUpperThreshold}()},
\code{\link{GNNMonSetValue}()},
\code{\link{GNNMonTakeData}()},
\code{\link{GNNMonTakeDistanceFunction}()},
\code{\link{GNNMonTakeKDTreeObject}()},
\code{\link{GNNMonTakeNearestNeighborDistances}()},
\code{\link{GNNMonTakeNumberOfNNs}()},
\code{\link{GNNMonTakeRadius}()},
\code{\link{GNNMonTakeRadiusFunction}()},
\code{\link{GNNMonTakeUpperThreshold}()},
\code{\link{GNNMonTakeValue}()}
}
\concept{Set/Take functions}
