% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonTakeValue}
\alias{GNNMonTakeValue}
\title{Take the value in a GNNMon object.}
\usage{
GNNMonTakeValue(gnnObj)
}
\arguments{
\item{gnnObj}{An GNNMon object.}
}
\value{
Just \code{gnnObj$Value}.
}
\description{
Takes the value from GNNMon monad object.
}
\seealso{
Other Set/Take functions: 
\code{\link{GNNMonSetData}()},
\code{\link{GNNMonSetDistanceFunction}()},
\code{\link{GNNMonSetKDTreeObject}()},
\code{\link{GNNMonSetLowerThreshold}()},
\code{\link{GNNMonSetNearestNeighborDistances}()},
\code{\link{GNNMonSetNumberOfNNs}()},
\code{\link{GNNMonSetRadius}()},
\code{\link{GNNMonSetRadiusFunction}()},
\code{\link{GNNMonSetUpperThreshold}()},
\code{\link{GNNMonSetValue}()},
\code{\link{GNNMonTakeData}()},
\code{\link{GNNMonTakeDistanceFunction}()},
\code{\link{GNNMonTakeKDTreeObject}()},
\code{\link{GNNMonTakeLowerThreshold}()},
\code{\link{GNNMonTakeNearestNeighborDistances}()},
\code{\link{GNNMonTakeNumberOfNNs}()},
\code{\link{GNNMonTakeRadius}()},
\code{\link{GNNMonTakeRadiusFunction}()},
\code{\link{GNNMonTakeUpperThreshold}()}
}
\concept{Set/Take functions}
