% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonUnit}
\alias{GNNMonUnit}
\title{Make a GNNMon Unit}
\usage{
GNNMonUnit(data)
}
\value{
An S3 class "GNNMon". In other words, a list with the attribute "class" set to "GNNMon".
}
\description{
Creates a monad object.
}
