% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonClassify}
\alias{GNNMonClassify}
\title{Classify for "membership".}
\usage{
GNNMonClassify(gnnObj, points)
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{points}{A numeric vector (a point) or a matrix with
number of columns that equals the dimension of monad's points.
(I.e. \code{ncol(gnnObj$Data)}).}
}
\value{
A GNNMon object
}
\description{
Classifies a given point as close enough to
the monad points (TRUE) or too distant (FALSE).
}
\details{
The result is assigned to \code{gnnObj$Value}.
}
