% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonComputeMatrixDistances}
\alias{GNNMonComputeMatrixDistances}
\title{Compute matrix-vector distances.}
\usage{
GNNMonComputeMatrixDistances(gnnObj, point, distanceMethod = "euclidean")
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{point}{A numeric vector.}

\item{distanceMethod}{A string for the distance method.
One of "euclidean", "cosine".}
}
\value{
A GNNMon object
}
\description{
Computes the distances of a point to the rows of monad's
data matrix using a specified method.
}
\details{
The result is assigned to \code{gnnObj$Value}.
}
