% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonComputeNearestNeighborDistances}
\alias{GNNMonComputeNearestNeighborDistances}
\title{Compute nearest neighbors distances.}
\usage{
GNNMonComputeNearestNeighborDistances(gnnObj, nTopNNs = 6,
  method = "euclidean")
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{nTopNNs}{Number of top nearest neighbors.}

\item{method}{A string for the distance method.
One of "euclidean", "cosine".}
}
\value{
A GNNMon object
}
\description{
Computes the nearest neighbors and corresponding distances
using a specified method for each row of monad's data matrix.
}
