% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonComputeThresholds}
\alias{GNNMonComputeThresholds}
\title{Compute proximity radius thresholds.}
\usage{
GNNMonComputeThresholds(
  gnnObj,
  nnsRadiusFunction = mean,
  thresholdsIdentifier = OutlierIdentifiers::HampelIdentifierParameters
)
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{nnsRadiusFunction}{The function used to compute global
nearest neighbors radius.}

\item{thresholdsIdentifier}{An outliers identification parameters function
that computes to hyper-sphere thresholds from the NN hyper-spheres radiuses.}
}
\value{
A GNNMon object
}
\description{
Estimates the upper threshold of proximity radius.
}
\details{

}
