% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonFindNearest}
\alias{GNNMonFindNearest}
\title{Find nearest points.}
\usage{
GNNMonFindNearest(
  gnnObj,
  point,
  n = 12,
  radius = Inf,
  method = "scan",
  distanceFunction = "euclidean"
)
}
\arguments{
\item{gnnObj}{A GNNMon object}

\item{point}{A numeric vector.}

\item{n}{Number of nearest neighbors.}

\item{radius}{Radius within which the nearest neighbors are taken.}

\item{method}{A string for the nearest neighbors finding method.
One of "scan", or  "kdtree".}

\item{distanceFunction}{A string for the distance function.
One of "euclidean", "cosine".}
}
\value{
A GNNMon object
}
\description{
Finds the nearest points from the monad to a given point.
}
\details{
The result is assigned to \code{gnnObj$Value}.
This is a fairly brute force algorithm because in order to find
the top nearest neighbors the distances to all monad points are computed.
}
