% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonOption}
\alias{GNNMonOption}
\title{Optional function application to monad's object.}
\usage{
GNNMonOption(gnnObj, f)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{f}{A function to be applied to the monad object.}
}
\value{
A GNNMon object.
}
\description{

}
\details{
In general \code{f} should return a monad object,
but that is not enforced.
}
