% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonSetNearestNeighborDistances}
\alias{GNNMonSetNearestNeighborDistances}
\title{Set NearestNeighborDistances.}
\usage{
GNNMonSetNearestNeighborDistances(gnnObj, NearestNeighborDistances)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{NearestNeighborDistances}{An object member to be set.}
}
\value{
An GNNMon object.
}
\description{
Sets NearestNeighborDistances into the monad object.
}
\seealso{
Other Set/Take functions: \code{\link{GNNMonSetData}},
  \code{\link{GNNMonSetDistanceMethod}},
  \code{\link{GNNMonSetLowerThreshold}},
  \code{\link{GNNMonSetNumberOfNNs}},
  \code{\link{GNNMonSetRadiusFunction}},
  \code{\link{GNNMonSetRadius}},
  \code{\link{GNNMonSetUpperThreshold}},
  \code{\link{GNNMonSetValue}},
  \code{\link{GNNMonTakeData}},
  \code{\link{GNNMonTakeDistanceMethod}},
  \code{\link{GNNMonTakeLowerThreshold}},
  \code{\link{GNNMonTakeNearestNeighborDistances}},
  \code{\link{GNNMonTakeNumberOfNNs}},
  \code{\link{GNNMonTakeRadiusFunction}},
  \code{\link{GNNMonTakeRadius}},
  \code{\link{GNNMonTakeUpperThreshold}},
  \code{\link{GNNMonTakeValue}}
}
\concept{Set/Take functions}
