% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonFunctions.R
\name{GNNMonSetValue}
\alias{GNNMonSetValue}
\title{Set the value in a GNNMon object.}
\usage{
GNNMonSetValue(gnnObj, value)
}
\arguments{
\item{gnnObj}{An GNNMon object.}

\item{value}{The new value.}
}
\value{
A GNNMon object.
}
\description{
Sets the value in a GNNMon monad object.
}
\details{
Assigns \code{value} to \code{gnnObj$Value}.
}
\seealso{
Other Set/Take functions: 
\code{\link{GNNMonSetData}()},
\code{\link{GNNMonSetDistanceFunction}()},
\code{\link{GNNMonSetKDTreeObject}()},
\code{\link{GNNMonSetLowerThreshold}()},
\code{\link{GNNMonSetNearestNeighborDistances}()},
\code{\link{GNNMonSetNumberOfNNs}()},
\code{\link{GNNMonSetRadius}()},
\code{\link{GNNMonSetRadiusFunction}()},
\code{\link{GNNMonSetUpperThreshold}()},
\code{\link{GNNMonTakeData}()},
\code{\link{GNNMonTakeDistanceFunction}()},
\code{\link{GNNMonTakeKDTreeObject}()},
\code{\link{GNNMonTakeLowerThreshold}()},
\code{\link{GNNMonTakeNearestNeighborDistances}()},
\code{\link{GNNMonTakeNumberOfNNs}()},
\code{\link{GNNMonTakeRadius}()},
\code{\link{GNNMonTakeRadiusFunction}()},
\code{\link{GNNMonTakeUpperThreshold}()},
\code{\link{GNNMonTakeValue}()}
}
\concept{Set/Take functions}
