% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNNMonCore.R
\name{GNNMonTakeValue}
\alias{GNNMonTakeValue}
\title{Take the value in a GNNMon object.}
\usage{
GNNMonTakeValue(gnnObj)
}
\arguments{
\item{gnnObj}{An GNNMon object.}
}
\value{
Just \code{gnnObj$Value}.
}
\description{
Takes the value from GNNMon monad object.
}
\seealso{
Other Set/Take functions: \code{\link{GNNMonSetData}},
  \code{\link{GNNMonSetDistanceMethod}},
  \code{\link{GNNMonSetLowerThreshold}},
  \code{\link{GNNMonSetNumberOfNNs}},
  \code{\link{GNNMonSetRadiusFunction}},
  \code{\link{GNNMonSetRadius}},
  \code{\link{GNNMonSetUpperThreshold}},
  \code{\link{GNNMonSetValue}},
  \code{\link{GNNMonTakeData}},
  \code{\link{GNNMonTakeDistanceMethod}},
  \code{\link{GNNMonTakeLowerThreshold}},
  \code{\link{GNNMonTakeNumberOfNNs}},
  \code{\link{GNNMonTakeRadiusFunction}},
  \code{\link{GNNMonTakeRadius}},
  \code{\link{GNNMonTakeUpperThreshold}}
}
\concept{Set/Take functions}
